requirements.txt
click==8.2.1
numpy==2.3.1
opencv-python==4.11.0.86
pillow==11.3.0
ExifRead